feat(about-page): generalize button component with <a> tags and clean up buttons
Generalized the Button component to use <a> tags and applied the changes across the site. Updated the buttons in the About page to link properly, setting the href to # for Event and Project pages for now.
