+++ src/app.html
@@ -1,5 +1,5 @@
 <!doctype html>
-<html lang="en">
+<html lang="en" class="scroll-smooth">
     <head>
         <title>UP CSI - Center for Student Innovations</title>
         <link rel="icon" href="%sveltekit.assets%/favicon.png" />
