--- src/lib/components/Button.svelte
@@ -1,15 +1,0 @@
-<script lang="ts">
-    import type { Snippet } from 'svelte';
-
-    interface Props {
-        children?: Snippet;
-    }
-
-    const { children }: Props = $props();
-</script>
-
-<button
-    class="rounded-full bg-black px-4 py-3 text-center font-inter text-csi-white dark:bg-csi-blue dark:text-csi-black"
->
-    {@render children?.()}
-</button>
