+++ src/lib/components/LinkButton.svelte
@@ -3,13 +3,15 @@
 
     interface Props {
         children?: Snippet;
+        href?: string;
+        class?: string;
     }
 
-    const { children }: Props = $props();
+    const { children, href = '/#' }: Props = $props();
 </script>
 
 <a
-    href="mailto:lino@up-csi.org"
+    {href}
     class="rounded-full bg-black px-4 py-3 text-center font-inter text-csi-white no-underline dark:bg-csi-blue dark:text-csi-black"
 >
     {@render children?.()}
