+++ src/routes/about/MissionVision.svelte
@@ -10,6 +10,7 @@
 
 <section
     class="flex flex-col gap-16 rounded-3xl bg-csi-neutral-50 px-8 py-12 font-dm text-csi-black md:px-12 dark:bg-csi-neutral-950 dark:text-csi-white"
+    id="mission-vision"
 >
     <div class="flex flex-col gap-8">
         <div class="text-center text-2xl">Mission</div>
